fixed version numbers for setup.py
