--- plantgw/version.py
@@ -1,3 +1,0 @@
-"""Plantgateway version number."""
-
-__version__ = '0.5.0'
