+++ setup.py
@@ -8,23 +8,19 @@
 ##############################################
 """Setup for plantgateway."""
 
-import sys
 from setuptools import setup
-from plantgw.version import __version__ as version
+from plantgw import __version__
 
 INSTALL_REQUIRES = ['bluepy==1.1.4', "paho-mqtt", 'pyyaml', "miflora==0.4"]
-if sys.version_info < (3, 0):
-    INSTALL_REQUIRES.append('mock')
-
 
 setup(
     name='plantgateway',
-    version=version,
+    version=__version__,
     description='Bluetooth to mqtt gateway for Xiaomi Mi plant sensors',
     author='Christian Kühnel',
     author_email='christian.kuehnel@gmail.com',
     url='https://www.python.org/sigs/distutils-sig/',
     packages=['plantgw'],
     install_requires=INSTALL_REQUIRES,
-    scripts=['plantgw/plantgateway'],
+    scripts=['plantgateway'],
     )
